third_party/blink/web_tests/http/tests/misc/onload-detach-during-csp-frame-src-none-expected.txt

CONSOLE ERROR: line 11: Refused to frame '' because it violates the following Content Security Policy directive: "frame-src 'none'".

PASS
